Added mockups for different views
Most are self explanatory. The weird red circle on the List view are
supposed to be icons that represent each topic, but I don’t have those
yet. The weird blue layout for the edit screen is from the Android
Calendar app. Its basically the idea I had in mind, but I didn’t have
time to mock up a specific one.